Fix relative paths access for mp3 files
They we're currently being accessed like ./file-name.mp3 but
that was relative to the directory the program was called in.
Instead it needs to be relative to the directory the file is in.

We can achieve this with the __dirname variable which gives us the
directory path of the current firectory that the file resides in.